Read files from a remote samba directory using a RegEx
The function listMyFiles takes a remote samba directory and a file extension
to list all files that satisfy the regular expression.  The problem was
how to get files listed on the remote server.  I could've found a solution faster
if the requirements were to list files from a locally mounted samba share, but I'm trying
to avoid mounting to access the fotos.  Listing them directly into a data structure
seems to be a good choice.