listFiles function signature changed to include a regEx
The samba-clientby eFlex Systems engineers had to changed for my purposes.
Its file signature only required a file name prefix, and a file name suffix.
The body of its version of listFiles helped my discover why the 'dir' command
was not listing the directory contents.  The reason was because the samba client
needed to have a wildcard * after the directory path to list all content.