Update README with a new TODOS to handle the samba asset share
A samba share that holds the media assets must be remotely accessed
to stream data to app.  To setup the app, all assets must be read to
memory.  The utility listFiles will help to get the stream started.

First, regular expressions must be created to handle all the different
naming conventions of the files.  The differences in namng conventions
come from having taken photos from both a Samsung Galaxy phone and
a Google Pixel 2XL.