Re-write files with no nominal file size from first iteration of getFotos.js
Initially, getFotos.writeP function would make a network request to copy
all files from Samba NAS to local machine asynchronously.  This network request
would overload the network causing some files to not be copied.  For those
that were not copied to local machine, the file name existed with a zero file
size.  This utility function was written to solve the problem of writing those
missed files to the local machine.

Moving forward, we will not make a heavy and bloated network request using
Promise.all, nor with Promise.allSettled, nor with Promise.race.  Instead, we
will chain the promises so that it does not overload the network.  How we will
chain is done in reDoNoSizePhotos.js.